island: move core/connection/gallerySync stores in + group hatches under toolkit.stores
Continues the device-store consolidation.

- core / connection / gallerySync stores -> modules/island/src/stores/* via the
  shim pattern (the @/stores/* files become 1-line re-exports; consumers unchanged).
- Two coupled types relocated into island so the stores are standalone-build-safe:
  WebSocketStatus (was @/services/ws-types) and PhotoInfo (was @/types/asg). Both old
  homes now re-export from island, so their other importers are unchanged.
- Grouped the raw-store escape hatches under toolkit.stores.{glasses,display,core,
  connection,gallerySync} (was toolkit.glassesStore/displayStore — nothing consumed
  those yet, so zero-risk). Reminder: stores are the Mentra-app escape hatch, NOT the
  OEM contract; OEMs use the typed facades.
- Deferred cloudClientStatus: it imports cloud-v2 (@mentra/cloud-client) types, so it
  belongs with the cloud-client / session work, not this batch.
- jest: the @mentra/island mock serves the real stores via requireActual.

Verified: typecheck 0 new errors across all consumers; jest 265 pass.
