+++ mobile/modules/island/src/stores/connection.ts
@@ -1,0 +1,68 @@
+import {create} from "zustand"
+
+// Relocated from the host @/services/ws-types so this store is self-contained
+// inside island. The host @/services/ws-types now re-exports this (a shim), so its
+// other importers are unchanged.
+export enum WebSocketStatus {
+  DISCONNECTED = "disconnected",
+  CONNECTING = "connecting",
+  CONNECTED = "connected",
+  ERROR = "error",
+}
+
+interface ConnectionState {
+  status: WebSocketStatus
+  url: string | null
+  error: string | null
+  lastConnectedAt: Date | null
+  lastDisconnectedAt: Date | null
+  reconnectAttempts: number
+
+  setStatus: (status: WebSocketStatus) => void
+  setUrl: (url: string | null) => void
+  setError: (error: string | null) => void
+  incrementReconnectAttempts: () => void
+  resetReconnectAttempts: () => void
+  reset: () => void
+}
+
+export const useConnectionStore = create<ConnectionState>((set) => ({
+  status: WebSocketStatus.DISCONNECTED,
+  url: null,
+  error: null,
+  lastConnectedAt: null,
+  lastDisconnectedAt: null,
+  reconnectAttempts: 0,
+
+  setStatus: (status) =>
+    set((state) => ({
+      status,
+      error: status === WebSocketStatus.ERROR ? state.error : null,
+      lastConnectedAt: status === WebSocketStatus.CONNECTED ? new Date() : state.lastConnectedAt,
+      lastDisconnectedAt:
+        status === WebSocketStatus.DISCONNECTED || status === WebSocketStatus.ERROR
+          ? new Date()
+          : state.lastDisconnectedAt,
+    })),
+
+  setUrl: (url) => set({url}),
+
+  setError: (error) => set({error, status: WebSocketStatus.ERROR, lastDisconnectedAt: new Date()}),
+
+  incrementReconnectAttempts: () =>
+    set((state) => ({
+      reconnectAttempts: state.reconnectAttempts + 1,
+    })),
+
+  resetReconnectAttempts: () => set({reconnectAttempts: 0}),
+
+  reset: () =>
+    set({
+      status: WebSocketStatus.DISCONNECTED,
+      url: null,
+      error: null,
+      lastConnectedAt: null,
+      lastDisconnectedAt: null,
+      reconnectAttempts: 0,
+    }),
+}))
