+++ mobile/src/services/ws-types.ts
@@ -1,6 +1,3 @@
-export enum WebSocketStatus {
-  DISCONNECTED = "disconnected",
-  CONNECTING = "connecting",
-  CONNECTED = "connected",
-  ERROR = "error",
-}
+// WebSocketStatus moved into @mentra/island (with the connection store that uses it).
+// Re-exported here so existing `@/services/ws-types` importers stay unchanged.
+export {WebSocketStatus} from "@mentra/island"
